[cg] Add text module, move text stubs there
